fix(save-flow): the write-open redirect is a process fact, not a module's
Two things the live runs proved, `br-20260913-144813-963f` and
`br-20260913-145249-468d`.

The legal footer on the title screen was the one element the cover never
touched. `title_child_name_matches` carried the bare `CopyrightText`, the name
as written in `05_000_title.gfx`, but `TitleTopDialog`'s initializer asks two
different proxies: dialog+0xa48 for the bare `PressStart` (deobf-1.17.1
0x1409a9415) and the parent root for `Top/CopyrightText` (0x1409a98ac and
0x1409a991a, both `lea r8, [0x142b295a8]`). The bare spelling never matched an
argument, so the run hid `PressStart`, `Info`, `ProgressInfo` and
`Install_ProgressInfo` and nothing else.

Then the Save Game commit refused a save it could have made.
`save_dest_open_redirect_installed` asked `er-quit-menu-core`'s own
`SaveHookInstallState`, which had never installed anything because
`install_save_flow_game_task` is called only by the `er-save-game-row` shell;
the product ticks the stage machine from its own recurring task. Meanwhile
`er-quickload`'s save-redirect path hook held the `kernel32!CreateFileW` detour
and its body asks `save_dest_redirect_for_open` the same question, so the armed
window was being read the whole time. MinHook keys by target address and refuses
the second hook on one address, so a per-module answer is structurally wrong
here: it says no for whichever module lost. The predicate now asks the process,
through `er_save_redirect::core_createfilew_installed_in_process`.

Measured after the change: redirect armed live `ER0000.co2` -> target
`ER0000.sl2`, two write-opens redirected, destination verified (bnd4, length and
structure all ok), `DEGRADED COMMIT COMPLETE`.

+++ crates/er-quit-menu-core/src/save_picker_dinput_nav.rs
@@ -144,7 +144,7 @@ unsafe extern "system" fn dinput_kb_get_state_hook(
     }
     let call: UnionFn = unsafe { std::mem::transmute::<usize, UnionFn>(next) };
     let raw = unsafe { call(device, size, data, unused) };
-    let (hr, size, data) = (raw as i32, size as usize, data as *const u8);
+    let (hr, size, data) = (raw as i32, size, data as *const u8);
     if hr != 0 || data.is_null() || size <= DINPUT_KEYBOARD_BUFFER_FLOOR {
         return raw;
     }
